Module forest_libp2p::chain_exchange

Structs

ChainExchangeBehaviour

ChainExchangeProtocolName

Type to satisfy ProtocolName interface for ChainExchange RPC.

ChainExchangeRequest

The payload that gets sent to another node to request for blocks and messages.

ChainExchangeResponse

The response to a ChainExchange request.

CompactedMessages

Contains all BLS and SECP messages and their indexes per block

TipsetBundle

Contains the blocks and messages in a particular tipset

Enums

ChainExchangeResponseStatus

Status codes of a chain_exchange response.

Constants

CHAIN_XCHG_PROTOCOL_ID

Libp2p protocol ID for ChainExchange.

HEADERS

ChainExchange Filecoin header set bit.

MESSAGES

ChainExchange Filecoin messages set bit.

Functions

make_chain_exchange_response

Builds chain exchange response out of chain data.

Type Definitions

ChainExchangeCodec

ChainExchange protocol codec to be used within the RPC service.