Module forest_libp2p::rpc

Structs

InboundCodec
Codec used for inbound connections. Decodes the inbound message into a RPCRequest, and encodes the RPCResponse to send.

OutboundCodec
Codec used for outbound connections. Encodes the outbound message into a RPCRequest to send, and decodes the RPCResponse when received.

RPC
The RPC behaviour that gets consumed by the Swarm.

RPCHandler

Enums

RPCError

RPCEvent
The return type used in the behaviour and the resultant event from the protocols handler.

RPCMessage
Messages sent to the user from the RPC protocol.

RPCRequest
RPCRequest payloads for request/response calls

RPCResponse
RPCResponse payloads for request/response calls

Constants

RESPONSE_TIMEOUT
The time (in seconds) before a substream that is awaiting a response from the user times out.

Type Definitions

RequestId