Crate forest_libp2p

Modules

chain_exchange

hello

rpc

Structs

Libp2pConfig
Libp2p config for the Forest node.

Libp2pService
The Libp2pService listens to events from the Libp2p swarm.

Multiaddr
Representation of a Multiaddr.

PeerId
Identifier of a peer of the network.

Topic
A gossipsub topic.

Enums

NetRPCMethods
Network RPC API methods used to gather data from libp2p node.

NetworkEvent
Events emitted by this Service.

NetworkMessage
Messages into the service to handle.

Protocol
Protocol describes all possible multiaddress protocols.

PubsubMessage
Message types that can come over GossipSub

Constants

PUBSUB_BLOCK_STR
Gossipsub Filecoin blocks topic identifier.

PUBSUB_MSG_STR
Gossipsub Filecoin messages topic identifier.

Functions

build_transport
Builds the transport stack that LibP2P will communicate over.

get_keypair
Fetch keypair from disk, returning none if it cannot be decoded.

Type Definitions

IdentTopic

Multihash
A Multihash with the same allocated size as the Multihashes produces by this derive.