Crate forest_libp2p

Modules

chain_exchange

ed25519

Ed25519 keys.

hello

rpc

Structs

Libp2pConfig

Libp2p configuration for the Forest node.

Libp2pService

The Libp2pService listens to events from the libp2p swarm.

Multiaddr

Representation of a Multiaddr.

PeerId

Identifier of a peer of the network.

Topic

A gossipsub topic.

Enums

Keypair

Identity keypair of a node.

NetRPCMethods

Network RPC API methods used to gather data from libp2p node.

NetworkEvent

Events emitted by this Service.

NetworkMessage

Messages into the service to handle.

Protocol

Protocol describes all possible multiaddress protocols.

PubsubMessage

Message types that can come over GossipSub

Constants

PUBSUB_BLOCK_STR

Gossipsub Filecoin blocks topic identifier.

PUBSUB_MSG_STR

Gossipsub Filecoin messages topic identifier.

Functions

build_transport

Builds the transport stack that libp2p will communicate over.

get_keypair

Fetch key-pair from disk, returning none if it cannot be decoded.

Type Definitions

IdentTopic

Multihash

A Multihash with the same allocated size as the Multihashes produces by this derive.

Derive Macros

Multihash